Fix a bug with overwriting part of the conf

+++ include/n2n_typedefs.h
@@ -492,7 +492,7 @@ typedef struct n2n_edge_conf {
     uint32_t register_ttl;                           /**< TTL for registration packet when UDP NAT hole punching through supernode. */
     struct sockaddr *bind_address;                   /**< The address to bind to if provided */
     n2n_sock_t preferred_sock;                       /**< propagated local sock for better p2p in LAN (-e) */
-    uint16_t mgmt_port;
+    uint32_t mgmt_port;     // TODO: ports are actually uint16_t
     bool connect_tcp;                                /** connection to supernode 0 = UDP; 1 = TCP */
     uint8_t sn_selection_strategy;                  /**< encodes currently chosen supernode selection strategy. */
     n2n_auth_t auth;
